Move audio_recorder code into alsa module
