+++ Makefile
@@ -3,7 +3,7 @@ SD=sdb
 
 LIBS=-lc -lasound -lsndfile -lm
 
-bin/app: package/src/app.c bin/display.o bin/alsa.o bin/metronome.o bin/input.o bin/menu.o bin/tone.o bin/libfft.o bin/audio_recorder.o
+bin/app: package/src/app.c bin/display.o bin/alsa.o bin/metronome.o bin/input.o bin/menu.o bin/tone.o bin/libfft.o
 	$(CC) $(CFLAGS) $^ -o $@ $(LIBS)
 
 bin/%.o: package/src/%.c
@@ -38,7 +38,7 @@ bin/display_test: package/src/display_test.c bin/display.o
 bin/tone_test: package/src/tone_test.c bin/tone.o bin/libfft.o
 	$(CC) $(CFLAGS) $^ -o $@ -lm
 
-bin/record_test: package/src/record_test.c bin/tone.o bin/libfft.o bin/audio_recorder.o bin/alsa.o
+bin/record_test: package/src/record_test.c bin/tone.o bin/libfft.o bin/alsa.o
 	$(CC) $(CFLAGS) $^ -o $@ -lm -lasound -lsndfile
 
 display_test: bin/display_test
