--- bbb/meta-bbb/recipes-core/trebletrouble/files/src/audio_recorder.h
@@ -1,5 +1,0 @@
-#include <stdint.h>
-#include <stdlib.h>
-#include "tone.h"
-
-int recordWAV(Wave* wave, uint32_t duration);
